+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Changes:
   * `github.com/hashicorp/vault/sdk` v0.9.2 => v0.10.2
   * `github.com/prometheus/client_golang` v1.16.0 => v1.17.0
 
+Improvements:
+* Injector can set CA certificate for injected pods via `AGENT_INJECT_VAULT_CACERT_BYTES` env var or `-vault-cacert-bytes` flag [GH-507](https://github.com/hashicorp/vault-k8s/pull/507)
+
 ## 1.3.0 (August 16, 2023)
 
 Improvements: